moved layout components to dedicated layout folder
