+++ src/components/Logo/Logo.js
@@ -1,0 +1,11 @@
+import React from "react"
+import { StyledLogo, TitleText, SubTitleText } from "./Logo.styled"
+
+const Logo = () => (
+  <StyledLogo>
+    <TitleText>Filipe Estácio</TitleText>
+    <SubTitleText>A Civil Engineer gone Digital</SubTitleText>
+  </StyledLogo>
+)
+
+export default Logo
