+++ src/layout/Header/Header.js
@@ -1,5 +1,5 @@
 import React from "react"
-import Logo from "../logo"
+import Logo from "../../components/Logo"
 import { StyledHeader } from "./Header.styled"
 
 const Header = () => (
