+++ src/pages/skills.js
@@ -1,5 +1,5 @@
 import React from "react"
-import Layout from "../components/layout"
+import Layout from "../layout"
 
 const Skills = () => {
   return (
